+++ test/fixtures/index.js
@@ -1,10 +1,23 @@
 const fs = require('fs'); // import syntax doesn't work inside karma
 
-const fixtures = {
-	'libraries.json': fs.readFileSync(__dirname + '/libraries.json'),
-	'library.json': fs.readFileSync(__dirname + '/library.json'),
-	'libraryVersions.json': fs.readFileSync(__dirname + '/libraryVersions.json'),
-};
+const fixtureNames = [
+	'libraries.json',
+	'library.json',
+	'libraryVersions.json',
+	'test-library-publish-0.0.1.tar.gz',
+	'test-library-publish-0.0.2.tar.gz'
+];
+
+function readFixtures(fixtureNames) {
+	const fixtures = {};
+	for (let idx in fixtureNames) {
+		const name = fixtureNames[idx];
+		fixtures[name] = fs.readFileSync(`${__dirname}/${name}`);
+	}
+	return fixtures;
+}
+
+const fixtures = readFixtures(fixtureNames);
 
 function read(filename) {
 	if (!fixtures[filename]) {